Make a stack of companion graphs from shared drive data

Make a stack of companion graphs from shared drive data

Usage

driveplot_companions(
  shareddata,
  x,
  ys,
  colorvar = NULL,
  xlabel = NULL,
  ylabels = NULL,
  colorpalette = "#03F",
  showlegend = TRUE,
  legendtitle = NULL,
  spacing = 0.05,
  plotheight = "98vh"
)

Arguments

shareddata

A SharedData object containing observations to be plotted.

x

The variable from shareddata to be plotted on the horizontal axis.

ys

A vector or list of 1 to 4 variables from shareddata to be plotted on the vertical axes of the companion graphs.

colorvar

The variable in shareddata to which color should be mapped.

xlabel

The label for the variable on the horizontal axis.

ylabels

A vector or list of 1 to 4 labels for the vertical axes of the companion graphs.

colorpalette

The color palette for the plot map; either a single color (e.g., "red") or one of "viridis", "inferno", "magma", or "plasma".

showlegend

Show the plot legend (TRUE) or not (FALSE).

legendtitle

The title for the plot legend.

spacing

A value between 0 and 1 for the space between the companion graphs.

plotheight

The height of the stack of companion graphs in CSS units, e.g, "98vh"

Value

A stack of plotly scatterplots.

Examples

library(crosstalk)
data(drive7)
shared_drive <- SharedData$new(drive7)

# Linked time series of speed, headings (in GPS and gyro), and GPS quality
driveplot_companions(
     shareddata = shared_drive,
     x = time_cst,
     ys = c(speed_mph, gyro_heading, gps_heading),
     colorvar = gps_pdop,
     xlabel = "Time",
     ylabels = c("Speed (mph)", "Gyro Heading (degrees)",
                 "GPS Heading (degrees)"),
     colorpalette = "viridis")