Blackbox-Driving Data for a number of drives

This dataset contains 28 drives ranging in length from just over 2 hours to just 45 secs.

Usage

nds_data

Format

A tibble with 23187 rows and 37 columns

drive

integer, identifier for a drive; a drive is defined as the time between starting the car and turning it off again.

time_utc

POSIXct time stamp for each observation in UTC

time_cst

POSIXct time stamp for each observation in Central time

GPS based measurements for each observation:

gps_long

numeric, geographic longitude

gps_lat

numeric, geographic latitude

gps_elevation

numeric, elevation above sea-level (in m)

gps_heading

numeric, direction of the car (in degrees)

gps_fix

numeric

gps_sats

integer, the number of satellites covering a position.

gps_pdop

numeric, total dilution of precision

gps_hdop

numeric, horizontal dilution of precision

gps_vdop

numeric, vertical dilution of precision

speed_mph

numeric, speed of the vehicle (in miles per hour)

speed_source

character, one of "gps" or "obd"

distance_miles

numeric, distance travelled during a drive (in miles)

Measurements by the on-board gyroscope:

accel_x

numeric, forward acceleration (in g)

accel_y

numeric, sidewards acceleration (in g)

accel_z

numeric, upwards acceleration (in g)

accel_event

binary, did an acceleration event occur? 0/1

accel_event_cat

character, categorization of acceleration event in low, medium, and high

gyro_x

numeric, angular rotation rate about the right-to-left axis (in degrees/sec)

gyro_y

numeric, angular rotation rate about the forward-to-backward axis.

gyro_z

numeric, angular rotation rate about the up-to-down axis.

gyro_heading

numeric, calculated heading utilizing compass data and gyro data (in degrees). Note that we use non-standard congruence classes for continuity of consecutive values.

grav_x

numeric, lateral gravity (in g)

grav_y

numeric, longitudinal gravity

grav_z

numeric, gravity (up-down)

roll

numeric, angle about the forward-to-backward axis. 0 = level, positive = tilted to the left

pitch

numeric, angle about the left-to-right axis. 0 = level, positive = tilted downward (i.e., vehicle going downhill)

engine_rpm

numeric, current engine revolutions per minute value

engine_throttle

numeric, percentage of engine throttle (0-100)

cumulative_drive_dist_mi

numeric

gps_minute

factor

gps_heading_raw

numeric, gps heading modulo 360

gyro_heading_raw

numeric, gyro heading modulo 360